make src dir and make datasheet path dynamic
